Removed all DEVELOPER_SUPPORT #ifdef's from the code.  The methods will now be compiled into the core.  It's up to the GUI's to call (or not call) the DEVELOPER methods.
Changed the behaviour of Console::saveProperties() in that it now accepts a
boolean variable 'merge', which if true, will make a call to
PropertiesSet::merge() and indicate that these properties are to be saved
into a stella.pro file.

If merge is false, it simply saves the properties to the specified file
(as before).