Added a new VC7 solution called 'Stella_Emulator' which contains the following two projects:
Stella  - the commandline Stella application
StellaX - the Windows GUI frontend

As of now, Stella compiles *without* the snapshot and OpenGL options
and works fine.  Adding these options will be as simple as finding
the required libs/include files and activating the relevant DEFINE's.

StellaX compiles and runs, but it doesn't do much ATM.  There is still
some amount of work to do in this area.

Since the Linux version is completely finished, and the MacOSX version
just needs some minor updates, most of my time will be spent getting the
StellaX GUI up to speed.  Then release 1.4 will be ready ...