Moved the main processing loop out of mainSDL and into OSystem::mainLoop(). Also, moved creation of new console into OSystem::createConsole.
Added hooks for browser mode (coming soon).

Added Ctrl-r to reload the currently loaded ROM.  It has the nice side
effect of testing whether another ROM can be reloaded while the emulation
is running.  Since this was successful, adding the ROM browser mode will
be very easy.

Added ability to actually change the framerate from the VideoDialog.

Reworked FrameBuffer::setPalette to be called only when a new console
is created (and we know that MediaSource::palette() is defined).

When toggling between NTSC/PAL mode, the video framerate is now toggled
between 50 and 60 (still TODO is change the audio framerate).