Added TabWidget and modified LauncherDialog to use it.
Changed many conflicting instances (in overloaded methods) of uInt32, uInt16,
Int32 and Int16 to just plain 'int'.  After spending two hours trying to find
out why a method wasn't being called, only to see it was because of conflicting
definitions of 'int', I said to hell with it and removed all instances of
those portability functions from the GUI stuff (thank Brian for the advice).

Eventually, they'll be removed from the whole codebase.  Already we're running
into problems with them on 64-bit machines.