Watch support: you can set up to 10 arbitrary expressions which will get evaluated and printed before each prompt. Since they're evaluated at print-time, you can say e.g. "watch a" or "watch *myCounter" and it'll do what you want.
Added "dump" command. This dumps any address you give it, not just RAM,
and it respects the current base setting (bin, dec, hex). Actually I may
rename the "ram" command to "mem" and allow it to work on non-RAM, and
merge the "dump" command with that.

The debugger state() now respects the default base (the registers are
printed in whatever base the user has selected with the "base" command).