Added keyboard selection and navigation to checkbuttons (use space or return key to toggle a selected checkbox).  Generally cleaned up the CheckboxWidget class, so that it centers vertically based on box/font size, and correctly determines its own size.
Added all P0/P1 info from 'tia' command to the TIA tab.  Some work
is still required on the NUSIZx registers, though.  I expect the
remainder of the TIA tab will be completed quickly, now that most
of the infrastructure is in place.  Still TODO is add methods to
TIADebug to enable/disable collisions.