Made all statefiles be named according to their romname as given in the properties file (instead of using md5sum).  The md5sum naming is no longer needed, and made it very difficult for users to distribute the state files.
Similar to above, removed ability to name snapshots according to
md5sum, and instead use romname.  Consequently, removed all
references to 'ssname' commandline argument and GUI related
config options.

Added ability to specify a different 'base directory' for settings,
state files, etc.  Normally, this is chosen at compile time, and is
set to $HOME/.stella/ for Linux and OSX, and the current dir for Win32.
Now, if you set the 'STELLA_BASEDIR' environment variable, it will
use that path instead.  Note that this must be set every time Stella
is run, since there's no way to save it to a settings file (since
the settings file location is dependent on what the env var contains).