Made the error message clearer when a snapshot file couldn't be created.
Made the default snapshot folder be $BASEDIR/snapshots, and got rid
of the previous default './'.  I suspect this relative pathname was
causing a problem.  The user is still free to choose another
snapshot dir, but the one in $BASEDIR is always created anyway.

Removed the logic that entering Stella directly from a ROM wouldn't
allow one to enter the ROM launcher afterwords.  You can now always
exit from a ROM back to the launcher, no matter how Stella was
launched.