Added more sophisticated framerate calculation based on the number of scanlines.  Previously, this was only done for the first few frames. However, that approach won't work for ROMs that change the number of scanlines later in their run.  Now, the framerate is automatically re-calculated at regular intervals within the TIA.
Reworked 'framerate' argument to override auto-frame calculation when
it contains a non-zero value, and use the given framerate instead.
Setting it to zero resumes auto-frame calculation.

Re-activated framerate option in the UI.  This can be changed while
the ROM is running, updates will be almost immediate.

Upped the release date, since more testing is required.