When starting Stella for the first time (ie, when the ROM directory hasn't been chosen), a dialog box now appears, asking the user to enter the default directory.  This directory will appear each time the ROM launcher is entered, until the user selects a directory.  This will hopefully eliminate bug reports from users that don't know what to do when first starting Stella.
Big changes to the manual, separating the necessary, up-front information
from the advanced stuff that can be considered later (or not at all).
Also added more links to ease navigation.  This should make the manual more
readable for those that have complained it's too long.

The setting for TV color type is now 'Color' and 'BW', instead of
'BlackAndWhite', as before.  The old one will continue to work, though.