Fixed bug when loading from ZIP files created in OSX.  In some cases, these archives contain directories named '__MACOSX' which contain link files that were being erroneously detected as actual ROMs.
Added FilesystemNode::isFile() method, and updated several places
in the code to use it.  Previously, determining whether something
was a file was simply testing if it wasn't a directory, but this
logic isn't always valid (it's possible to be neither a regular
file *nor* a directory).

Bumped version # for RC release.  It's getting very close ...