Temporary workaround for saving TIA image from debugger window; we only ever want to see the TIA, not the entire debugger area too.
This will be fixed for Stella 4.0, once I remove the @#*%$-ing software
rendering support.  That part of the code keeps coming back to haunt me
for years.  It will be *SO MUCH EASIER* to deal with these things when
everything is a FBSurface that can be stretched/scaled as necessary.